+++ ZitiDesktopEdge.Client/Properties/AssemblyInfo.cs
@@ -1,15 +1,16 @@
-﻿﻿using System.Reflection;
+﻿using System.Reflection;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("ziti-monitor")]
-[assembly: AssemblyDescription("Ziti Monitor and Update Service")]
+[assembly: AssemblyTitle("ZitiDesktopEdge.Client")]
+[assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("NetFoundry")]
-[assembly: AssemblyProduct("ZitiUpdateService")]
-[assembly: AssemblyCopyright("Copyright © NetFoundry 2020")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct("ZitiDesktopEdge.Client")]
+[assembly: AssemblyCopyright("Copyright ©  2020")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -19,7 +20,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("9ab81609-f5dd-40a6-87b5-f3279d48514d")]
+[assembly: Guid("26b30979-99b5-4102-bd0b-129bedff0057")]
 
 // Version information for an assembly consists of the following four values:
 //
@@ -31,5 +32,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.3.4.0")]
-[assembly: AssemblyFileVersion("1.3.4.0")]
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]