+++ service/cziti/dns-proxy.go
@@ -159,8 +159,11 @@ func runListener(ip *net.IP, port int, reqch chan dnsreq) {
 		server, err = net.ListenUDP(network, laddr)
 		if err == nil {
 			break
-		} else if attempts > maxAttempts{
+		} else if attempts > maxAttempts {
 			log.Panicf("An unexpected and unrecoverable error has occurred while %s: %v", "udp listening on network", err)
+		} else if attempts < (3 * maxAttempts / 4) {
+			// only log at debug until we hit 3/4 the max attempts to remove unnecessary warns from the log
+			log.Debugf("System not ready to listen on %v yet. Retrying after 500ms. This has happened %d times.", laddr, attempts)
 		} else {
 			log.Warnf("System not ready to listen on %v yet. Retrying after 500ms. This has happened %d times.", laddr, attempts)
 		}